requirements.txt
beautifulsoup4==4.11.1
Brotli==1.0.9
cachelib==0.9.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
dash==2.0.0
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-extensions==0.0.65
dash-html-components==2.0.0
dash-leaflet==0.1.23
dash-table==5.0.0
Deprecated==1.2.13
EditorConfig==0.12.3
Flask==2.1.3
Flask-Caching==2.0.1
Flask-Compress==1.12
gcloud==0.18.3
geobuf==1.1.1
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsbeautifier==1.14.6
jwcrypto==1.4.2
MarkupSafe==2.1.1
more-itertools==8.14.0
numpy==1.23.3
oauth2client==4.1.3
plotly==5.10.0
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-jwt==3.3.4
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
tenacity==8.0.1
urllib3==1.26.12
Werkzeug==2.0.0
wrapt==1.14.1
